requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
bs4==0.0.2
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
entrypoints==0.4
extra-streamlit-components==0.1.71
Faker==26.0.0
favicon==0.7.0
fonttools==4.53.1
gitdb==4.0.11
GitPython==3.1.43
google_search_results==2.4.2
h11==0.14.0
holehe==1.61
htbuilder==0.6.2
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lxml==5.2.2
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
more-itertools==10.3.0
murmurhash==1.0.10
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pdfminer==20191125
pillow==10.4.0
preshed==3.0.9
prometheus_client==0.20.0
protobuf==5.27.2
pyarrow==16.1.0
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pymdown-extensions==10.8.1
pyparsing==3.1.2
pypdf==4.2.0
PySocks==1.7.1
pyspellchecker==0.8.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
serpapi==0.1.5
setuptools==70.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
st-annotated-text==4.0.1
st-theme==1.2.3
streamlit==1.36.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.4.3
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-js-eval==0.1.7
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.4.2
termcolor==2.4.0
thinc==8.2.5
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
trio==0.26.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
validators==0.32.0
wasabi==1.1.3
watchdog==4.0.1
weasel==0.4.1
wrapt==1.16.0
WTForms==3.1.2
